# lib: recover panics within libraries and convert to informative errors