--- a/lib/errors.go
+++ b/lib/errors.go
@@ -19,9 +19,14 @@ package lib
 
 import (
 	"fmt"
+	"path"
+	"runtime"
+	"strings"
 
 	"github.com/google/cel-go/cel"
 	"github.com/google/cel-go/common"
+	"github.com/google/cel-go/common/types"
+	"github.com/google/cel-go/common/types/ref"
 )
 
 // DecoratedError implements error source location rendering.
@@ -80,5 +85,64 @@ func nodeID(err error) (id int64, ok bool) {
 		default:
 			return 0, false
 		}
+	}
+}
+
+type (
+	unop  = func(value ref.Val) ref.Val
+	binop = func(lhs ref.Val, rhs ref.Val) ref.Val
+	varop = func(values ...ref.Val) ref.Val
+
+	bindings interface {
+		unop | binop | varop
+	}
+)
+
+func catch[B bindings](binding B) B {
+	switch binding := any(binding).(type) {
+	case unop:
+		return any(func(arg ref.Val) (ret ref.Val) {
+			defer handlePanic(&ret)
+			return binding(arg)
+		}).(B)
+	case binop:
+		return any(func(arg0, arg1 ref.Val) (ret ref.Val) {
+			defer handlePanic(&ret)
+			return binding(arg0, arg1)
+		}).(B)
+	case varop:
+		return any(func(args ...ref.Val) (ret ref.Val) {
+			defer handlePanic(&ret)
+			return binding(args...)
+		}).(B)
+	default:
+		panic("unreachable")
+	}
+}
+
+func handlePanic(ret *ref.Val) {
+	switch r := recover().(type) {
+	case nil:
+		return
+	default:
+		// We'll only try 64 stack frames deep. There are a few recursive
+		// functions in extensions, but panic in those functions are unlikely
+		// or are already explicitly recovered as part of normal operation.
+		pc := make([]uintptr, 64)
+		n := runtime.Callers(2, pc)
+		cf := runtime.CallersFrames(pc[:n])
+		for {
+			f, more := cf.Next()
+			if !more {
+				break
+			}
+			file := f.File
+			if strings.Contains(file, "mito/lib") {
+				_, file, _ := strings.Cut(file, "mito/")
+				*ret = types.NewErr("%s: %s %s:%d", r, path.Base(f.Function), file, f.Line)
+				return
+			}
+		}
+		*ret = types.NewErr("%s", r)
 	}
 }