… lib/collections.go
… lib/file.go
… lib/http.go
… lib/json.go
… lib/regexp.go
… lib/xml.go